dask optimized slicing context
